+++ Magic/CHANGELOG.md
@@ -73,6 +73,7 @@
  - Added new builtin attributes: target_health, target_health_max, epoch
  - Add support for using saved disguises in LibsDisguises
  - Add ability to remove Gather friends
+ - Added unswappable wand option
 
 # 8.2.3